+++ R/download_worldclim_present.R
@@ -9,86 +9,87 @@
 #' @keywords internal
 
 download_worldclim_present <- function(dataset, bio_var, filename) {
-  # reset warnings for gdal to terra default
-  # this is necessary if sf was loaded in the mean time
-  terra::gdal(warn = 3)
-
-  # get resolution from the dataset name and convert it to the original
-  res_conversion <- data.frame(
-    our_res = c("10m", "5m", "2.5m", "0.5m"),
-    wc_res = c("10m", "5m", "2.5m", "30s")
-  )
-  wc_res <- res_conversion$wc_res[res_conversion$our_res == substr(dataset,
-    start = 15,
-    stop = nchar(dataset)
-  )]
-
+  # extract the version from the filename
+  version <- gsub("^.*_v([0-9]+\\.[0-9]+\\.[0-9]+)\\..*$", "\\1", filename)
+  # compose download paths
+  download_url <- filenames_worldclim_present(dataset=dataset,
+                                             bio_var = bio_var)
+  time_vector <- 1985
+  
+  # create band vectors
   # function to grab the number from the raster layer
   if (grepl("bio", bio_var)) {
-    postfix <- "bio.zip"
-    var_prefix <- "bio"
-  } else if (grepl("temperature_min", bio_var)) {
-    postfix <- "tmin.zip"
-    var_prefix <- "temperature_min_"
-  } else if (grepl("temperature_max", bio_var)) {
-    postfix <- "tmax.zip"
-    var_prefix <- "temperature_max_"
-  } else if (grepl("temperature_", bio_var)) {
-    postfix <- "tavg.zip"
-    var_prefix <- "temperature_"
-  } else if (grepl("precipitation_", bio_var)) {
-    postfix <- "prec.zip"
-    var_prefix <- "precipitation_"
+    band_vector <- paste0("bio", sprintf("%02d", 1:19))
   } else if (grepl("altitude", bio_var)) {
-    postfix <- "elev.zip"
-    var_prefix <- "elevation"
+    band_vector <- "altitude"
+  } else {
+    var_prefix <- substr(bio_var,1,nchar(bio_var)-2)
+    band_vector <- paste0(var_prefix, sprintf("%02d", 1:12))
+  }
+  # we are missing the special case for altitude
+  # elev.zip
+  #browser()
+  
+  # if we do not have a directory, create one
+  worldclim_dir <- file.path(get_data_path(),"worldclim_2.1")
+  if(!dir.exists(worldclim_dir)){
+    dir.create(worldclim_dir)
   }
 
-  base_url <- "https://biogeo.ucdavis.edu/data/worldclim/v2.1/base/wc2.1"
-  full_url <- paste(base_url, wc_res, postfix, sep = "_")
-  destfile <- tempfile()
-  # download this zip file into a temp file
-  curl::curl_download(full_url,
-    destfile = destfile,
+  # and now download the zip file (so the url is actually a local file)
+  worldclim_url <- file.path(worldclim_dir,basename(download_url))
+  curl::curl_download(download_url,
+    destfile = worldclim_url,
     quiet = FALSE
   )
-
-  # unzip it to a temporary directory
-  destpath <- file.path(tempdir(), "to_unzip")
-  utils::unzip(destfile, exdir = destpath)
-  wc_rast <- terra::rast(dir(destpath, pattern = ".tif", full.names = TRUE))
-  # sort out variable names
-  if (!(grepl("altitude", bio_var))) {
-    # digits at the end of the name are the key identifier of each variable
-    digits_at_end <- sprintf(
-      "%02d",
-      as.numeric(substr(
-        names(wc_rast),
-        regexpr("_\\d+\\b", names(wc_rast)) + 1,
-        nchar(names(wc_rast))
-      ))
-    )
-
-    # now we need to rename the layers
-    names(wc_rast) <-
-      paste0(var_prefix, digits_at_end)
-  } else {
-    names(wc_rast) <- "altitude"
+  # for all variables (except altitude) we match file names to bands based on the id at the end of the file name
+  files_in_zip <- utils::unzip(worldclim_url, list=TRUE)$Name
+  # TODO remove any readme from this list
+  files_in_zip <- files_in_zip[!files_in_zip %in% c("readme.txt")]
+  if (!grepl("altitude", bio_var)){
+    files_in_zip_id <- as.numeric(gsub("^.*_([0-9]+)\\.tif$", "\\1", files_in_zip))
+    if (!all(seq_len(length(band_vector) %in% files_in_zip_id))){
+      stop("some files are missing from the downloaded archive")
+    }    
+  } else { # for altitude
+    if (length(files_in_zip)!=1){
+      stop("we expected only one file in the elevation zip file")
+    }
   }
 
-  # and finally we save it as a netcdf file
-  time_bp(wc_rast) <- rep(35, nlyr(wc_rast))
+  # create a vrt for each variable
+  for (i in seq_len(length(band_vector))){
+    # build the vsizip paths
+    if (!grepl("altitude", bio_var)){
+      worldclim_vsizip <- paste0("/vsizip/",file.path(worldclim_url,files_in_zip[files_in_zip_id==i]))
+    } else { # special case for altitude
+      worldclim_vsizip <- paste0("/vsizip/",file.path(worldclim_url,files_in_zip[1]))
+    }
+    vrt_filename <- paste0(dataset,"_",band_vector[i],"_v",version,".vrt")
+    # create the vrt file
+    # vrt_path <- terra::vrt(x = worldclim_vsizip,
+    #                        filename = file.path(get_data_path(),vrt_filename),
+    #                        options="-separate", overwrite=TRUE, return_filename=TRUE)
+    vrt_path <- file.path(get_data_path(),vrt_filename)
+    sf::gdal_utils(
+      util = "buildvrt",
+      source = worldclim_vsizip,
+      destination = vrt_path,
+      options = c("-separate","-overwrite")
+    )
+    # edit the vrt metadata
+    edit_res <- vrt_set_meta(vrt_path = vrt_path, 
+                             description = band_vector[i],
+                             time_vector = time_vector,
+                             time_bp = FALSE)
+    if (!edit_res){
+      file.remove(vrt_path)
+      stop("something went wrong setting up this dataset", "\n the dataset will need downloading again")
+    }
+  }
+  if (!file.exists(filename)){
+    stop("something went wrong setting up this dataset", "\n the dataset will need downloading again")
+  }  
+  return(TRUE)
 
-  terra::writeCDF(wc_rast,
-    filename = filename, compression = 9,
-    split = TRUE, overwrite = TRUE
-  )
-
-  # fix time axis (this is a workaround if we open the file with sf)
-  nc_in <- ncdf4::nc_open(filename, write = TRUE)
-  ncdf4::ncatt_put(nc_in, varid = "time", attname = "axis", attval = "T")
-  ncdf4::nc_close(nc_in)
-
-  # clean up
-  unlink(file.path(destpath, "*"))
 }