+++ R/pastclim-package.R
@@ -1,0 +1,8 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+#' @import methods
+#' @import terra
+NULL